+++ packages/zhouhaifei-utils/.fatherrc.js
@@ -1,4 +1,0 @@
-export default {
-  entry: 'src/index.ts',
-  esm: { type: 'babel' },
-};