+++ packages/zhouhaifei-utils/src/index.ts
@@ -1,3 +1,0 @@
-export * from './locale';
-export * from './renderRoutes';
-export * from './serviceWorker';