+++ src/serviceWorker.ts
@@ -9,7 +9,6 @@
 
 /* To learn more about the benefits of this model and instructions on how to
    opt-in, read http://bit.ly/CRA-PWA */
-declare const process: any;
 const isLocalhost = Boolean(
   window.location.hostname === 'localhost' ||