+++ src/utils/browserHistory.ts
@@ -1,0 +1,3 @@
+import { createBrowserHistory, History } from 'history';
+
+export const browserHistory: History = createBrowserHistory({ basename: '/' });