+++ src/utils/index.ts
@@ -1,0 +1,4 @@
+export * from './browserHistory';
+export * from './locale';
+export * from './renderRoutes';
+export * from './request';